Move object_type.hpp to include
