+++ shards/core/runtime.hpp
@@ -18,11 +18,11 @@
 #include "pmr/unordered_map.hpp"
 #include "pmr/shared_temp_allocator.hpp"
 #include <shards/log/log.hpp>
+#include <shards/object_type.hpp>
 #include "shards_macros.hpp"
 #include "foundation.hpp"
 #include "inline.hpp"
 #include "utils.hpp"
-#include "object_type.hpp"
 #include "platform.hpp"
 
 #include <chrono>
