+++ shards/modules/gfx/shards_types.hpp
@@ -4,7 +4,7 @@
 #include "gfx.hpp"
 #include <shards/common_types.hpp>
 #include <shards/core/foundation.hpp>
-#include <shards/core/object_type.hpp>
+#include <shards/object_type.hpp>
 #include <shards/shards.hpp>
 #include <memory>
 #include <variant>
